feat: prevent pagination token re-use using context
In a service with multiple resource types, or a service used by
multiple users, it can be valuable to prevent malicious actors from
using pagination tokens outside of the context in which they are
intended to be used.

This change uses the `encryption_context` to bind pagination tokens
to the authenticated user and the API path for which they were
created. If someone else tries to use the token, or if a user tries
to take a token from one resource and use it in a different resource,
the decryption process will fail and the user request will be blocked.

In scenarios where the token encryption key is used across multiple
API Gateway instances, it would be valuable to include the API Gateway
ID in the context as well so that tokens cannot be taken across that
boundary. If the token encryption key is not shared, the additional
context of the API Gateway ID is unnecessary.